+++ ai-logic/firebase-ai/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 17.13.0
+
 - [feature] Expanded `SpeechConfig` to support `MultiSpeakerVoiceConfig` and `LanguageCode`.
   Integrated `SpeechConfig` into `GenerationConfig` to support Text-to-speech generation (TTS). (#8100)
 - [feature] Added `modelVersion` property to `GenerateContentResponse` (#8227)
@@ -183,3 +185,4 @@ using [specific Gemini models](/docs/vertex-ai/models).
 
 Note: This feature is in Public Preview, which means that it is not subject to any SLA or
 deprecation policy and could change in backwards-incompatible ways.
+